decision-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
fastapi==0.121.3
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.4.0
opencv-python==4.12.0.88
protobuf==5.29.5
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
